Switch to use tileDispatchUsingSCFFopOp
